requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2-binary==2.9.1
django-utils-six
django-model-utils==4.1.1
Pillow
django-filter
drf-yasg